requirements.txt
anthropic>=0.25.0
supabase>=2.4.0
streamlit>=1.35.0
apify-client>=1.6.0
python-docx>=1.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0
pandas>=2.2.0
